pyproject.toml
[tool.poetry]
name = "django-api-files-project"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
